refactor: split feature store modules and harden query parameters
