+++ src/Honua.Server/Features/FeatureServer/Models/Attachments/AddAttachmentResult.cs
@@ -1,0 +1,31 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+namespace Honua.Server.Features.FeatureServer.Models;
+
+/// <summary>
+/// Result of adding an attachment
+/// </summary>
+public sealed class AddAttachmentResult
+{
+    /// <summary>
+    /// Feature ID that owns the attachment
+    /// </summary>
+    public required long ObjectId { get; init; }
+
+    /// <summary>
+    /// Whether the operation was successful
+    /// </summary>
+    public required bool Success { get; init; }
+}
+
+/// <summary>
+/// Response for adding an attachment
+/// </summary>
+public sealed class AddAttachmentResponse
+{
+    /// <summary>
+    /// Add attachment result
+    /// </summary>
+    public required AddAttachmentResult AddAttachmentResult { get; init; }
+}
