+++ src/Honua.Server/Features/FeatureServer/Models/Relationships/QueryRelatedRecordsResponse.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+namespace Honua.Server.Features.FeatureServer.Models;
+
+/// <summary>
+/// Response model for queryRelatedRecords endpoint
+/// </summary>
+public sealed class QueryRelatedRecordsResponse
+{
+    /// <summary>
+    /// Array of related record groups, one per source object ID
+    /// </summary>
+    public required RelatedRecordGroup[] RelatedRecordGroups { get; init; }
+}
